# packages/server/api.yaml
swagger: '2.0'
info:
  description: SARA REST API definition
  version: 1.0.0
  title: SARA (Sales Reporter Applicaltion) REST services
  contact:
    email: miller.gonzalez@globant.com
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.

     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.

     The following syntax must be used in the 'Authorization' header :

         Bearer xxxxxx.yyyyyyy.zzzzzz
paths:
...
  /program:
    get:
      x-swagger-router-controller: program
      operationId: "getPrograms"
      summary: Returns the list of probrams.
      description: This will be used to populate the insert and update dropdowns
      security:
        - Bearer: []
      x-security-scopes:
        - admin
        - user
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/ProgramListResponse'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
    post:
      x-swagger-router-controller: program
      operationId: "addNewProgram"
      summary: Create a new program
      description: N/A
      security:
        - Bearer: []
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/ProgramBody'
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
...
  /program/{id}:
    get:
      x-swagger-router-controller: program
      operationId: "getProgramById"
      summary: get a program by id
      description: N/A
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          description: ID of program to fetch
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/ProgramResponse'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
    put:
      x-swagger-router-controller: program
      operationId: "updateProgram"
      summary: Update a program
      security:
        - Bearer: []
      description: N/A
      parameters:
        - in: path
          name: id
          description: ID of program to be updated
          required: true
          type: string
        - in: body
          name: body
          schema:
            $ref: '#/definitions/ProgramBody'
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
    delete:
      x-swagger-router-controller: program
      operationId: "deleteProgram"
      summary: Delete a program
      security:
        - Bearer: []
      description: N/A
      parameters:
        - in: path
          name: id
          description: ID of program to be deleted
          required: true
          type: string
        - in: body
          name: body
          schema:
            $ref: '#/definitions/ProgramBodyDelete'
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
...
  /programs:
    post:
      x-swagger-router-controller: program
      operationId: "getProgramsFiltered"
      summary: Returns the list of programs.
      description: This will be used to populate the insert and update dropdowns
      security:
        - Bearer: []
      x-security-scopes:
        - admin
        - user
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/FullProgramsBody'
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/ProgramListResponse'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
  /programs/{userId}:
    post:
      x-swagger-router-controller: program
      operationId: "getProgramsFull"
      summary: Returns the list of programs.
      security:
        - Bearer: []
      description: This will be used to populate the activity table
      parameters:
        - in: path
          name: userId
          description: ID of user to fetch programs
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/FullProgramsBody'
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/ProgramListResponse'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
  /programs/region/{regionId}:
    post:
      x-swagger-router-controller: program
      operationId: "getProgramsByRegionId"
      summary: Returns the list of programs.
      security:
        - Bearer: []
      description: This will be used to populate the activity table
      parameters:
        - in: path
          name: regionId
          description: ID of region to fetch programs
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/FullProgramsBody'
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/ProgramListResponse'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
...
  /activity:
    post:
      x-swagger-router-controller: activity
      operationId: "addNewActivity"
      summary: Create a new activity
      security:
        - Bearer: []
      description: N/A
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/ActivityBody'
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
    delete:
      x-swagger-router-controller: activity
      operationId: "deleteActivity"
      summary: Delete an activity
      security:
        - Bearer: []
      description: N/A
      parameters:
      - name: body
        in: body
        schema:
          $ref: '#/definitions/DeleteActivity'
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
  /activities/{userId}:
    post:
      x-swagger-router-controller: activity
      operationId: "getActivities"
      summary: Returns the list of activities by user.
      security:
        - Bearer: []
      description: This will be used to populate the activities table
      parameters:
        - in: path
          name: userId
          description: ID of user to fetch activities
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/GetActivitiesBody'
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/ProgramListResponse'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
  /activity/{id}:
    get:
      x-swagger-router-controller: activity
      operationId: "getActivityById"
      summary: get an activity by id
      security:
        - Bearer: []
      description: N/A
      parameters:
        - in: path
          name: id
          description: ID of activity to fetch
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/ActivityResponse'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'

    put:
      x-swagger-router-controller: activity
      operationId: "updateActivity"
      summary: Update an activity by id
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          description: ID of activity to update
          required: true
          type: string
        - name: body
          in: body
          schema:
            $ref: '#/definitions/ActivityBody'
      responses:
        200:
          description: Success
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
        default:
          description: Invalid request.
          schema:
            allOf:
              - $ref: '#/definitions/ResponseStatus'
              - $ref: '#/definitions/Error'
...
definitions:
  ActivityBody:
    description: Body used for creating a new activity.
    type: object
    properties:
      title:
        type: string
      campaignId:
        type: [string, 'null']
      formatId:
        type: string
      abstract:
        type: string
      regionId:
        type: string
      startDate:
        type: string
      endDate:
        type: string
      asset:
        type: [string, 'null']
      userId:
        type: string
      programId:
        type: string
      customerMarketing:
        type: [boolean, 'null']
  GetActivitiesBody:
    description: Body used for get the activities based in a date filter
    type: object
    properties:
      startDate:
        type: string
      endDate:
        type: string
  DeleteActivity:
    description: Body used for deleting one or more activities.
    type: object
    properties:
      activitiesId:
        type: array
        items:
          type: string
...
  ProgramBodyDelete:
    description: Body used for delete program.
    type: object
    properties:
      userId:
        type: string
  FullProgramsBody:
    description: Body used for retrieve full list of programs.
    type: object
    properties:
      programsStartDate:
        type: string
      programsEndDate:
        type: string
  ProgramBody:
    description: Body used for creating a new program.
    type: object
    properties:
      name:
        type: string
      owner:
        type: string
      budget:
        type: number
      metrics:
        type: number
      customerMessage:
        type: string
      regionId:
        type: string
      lifecycleStageId:
        type: array
        items:
          type: string
      apm1Id:
        type: array
        items:
          type: string
      apm2Id:
        type: array
        items:
          type: string
      industryId:
        type: array
        items:
          type: string
      segmentId:
        type: array
        items:
          type: string
      personaId:
        type: array
        items:
          type: string
      otherKpis:
        type: string
      year:
        type: number
      quarter:
        type: number
      year_quarter:
        type: number
...
  ProgramListResponse:
    description: A list of programs.
    type: object
    properties:
      result:
        type: array
        items:
          $ref: '#/definitions/Program'
  ProgramResponse:
    description: A program get by Id
    type: object
    properties:
      result:
        type: array
        items:
          $ref: '#/definitions/Program'
...
  ResponseStatus:
    type: object
    properties:
      info:
        description: The object containing the HTTP status for the response.
        type: object
        properties:
          status:
            description: The numeric HTTP status code for the response.
            type: integer
            example: Status HTTP Code
          code:
            description: The textual HTTP status code for the response.
            type: string
            example: OK
        required:
          - status
          - code
  ActivityResponse:
    description: Single activity
    type: object
    properties:
      result:
        $ref: '#/definitions/Activity'
...
  Error:
    properties:
      message:
        type: string
    required:
      - message

  #MODELS
  Activity:
    type: object
    properties:
      title:
        type: string
      campaignId:
        type: string
      tacticId:
        type: string
      formatId:
        type: string
      abstract:
        type: string
      regionId:
        type: string
      startDate:
        type: string
      endDate:
        type: string
      asset:
        type: string
      userId:
        type: string
      programId:
        type: string
      customerMarketing:
        type: boolean
  Program:
    type: object
    properties:
      program_id:
        type: string
      label:
        type: string
...
schemes: ['http', 'https']
consumes: ['application/json']
produces: ['application/json']
host: localhost:3000
basePath: /api/v1